Fix for uniqueId cache lookup on silent and interactive handlers
When there are two users in the cache with the same authority, clientId, tenant but different uniqueId's or displayableId's the cache will throw a mutliple token detected exception when acquiring a token silently or using an auto or never interactive client. The reason is because the userId and displayableId are not being considered during cache lookup. This change fixes that by considering those two attribute. The added test will fail when we do not consider the userId correctly. It passes with the changes.