Added an optional readonly address bar in Login Dialog Box.
Added a new optional DiplayAddressBar property which is 'False' by default in PlatformParameter and updated all dependent classes.
Wrapped CustomWebBrowser in UserControlWebBrowser along with a readonly TextBox that displays Adfs Address Url on top of the WebBrowser control.
It can assure the user that the login screen belongs to the ADFS.