Modified exception checking in tests
Added a helper method to check that the expected exceptions are being thrown and modified all of the tests to use the new helper.
Using the helper method is simpler and less error-prone; previously
some of the tests were not correctly checking for error and could have
been passing even if the expected exception was not thrown.
There are two cases in AuthenticationParametersTests where this is
happening (issue #611)