Update appveyor.yml to select NoWinRT sln
We are unable to build both MSBuild15 projects and WinRT projects at the
same time on App Veyor. This is because the image which contains VS2017
(for MSBuild 15) does not contain the tooling for WinRT (Windows 8.x).
On the flipside the image which contains the WinRT tooling does not
include VS2017/MSBuild 15.

For App Veyor builds we exclude the WinRT projects from the build by
using the ADAL.NET.NoWinRT.sln solution file.