Allow HttpClient to be provided when creating AuthenticationContext
When interacting with multiple different tenants new instances of the AuthenticationContext need to be created for each tenant.  Under the covers this causes a new instance of an HttpClient to be created.  This usage pattern causes connections to be re-established for each tenant and this can cause problems when working with large number of different tenants.

To resolve this, a new constructor has been added to the AuthenticationContext which allows a HttpClient instance to be provided.  This allows the user of the SDK to control the reuse of the HttpClient in scenarios where this is needed.